requirements.txt
# requerimientos
requests
gradio
python-dotenv
